refactor: keep next config thin
