+++ next.config.ts
@@ -1,37 +1,7 @@
 import type { NextConfig } from 'next'
 
+import { getImageRemotePatterns } from '@/lib/config/nextImageRemotePatterns'
 import { getSecurityHeaders as getRuntimeSecurityHeaders } from '@/lib/utils/securityHeaders'
-
-type ImageRemotePatterns = NonNullable<
-  NonNullable<NextConfig['images']>['remotePatterns']
->
-
-const SAFE_LOCAL_IMAGE_REMOTE_PATTERNS: ImageRemotePatterns = [
-  {
-    protocol: 'http',
-    hostname: 'localhost'
-  },
-  {
-    protocol: 'http',
-    hostname: '127.0.0.1'
-  },
-  {
-    protocol: 'http',
-    hostname: '[::1]'
-  }
-]
-
-const isDevelopment = () => process.env.NODE_ENV !== 'production'
-
-export const getImageRemotePatterns = (): ImageRemotePatterns => {
-  const patterns: ImageRemotePatterns = [{ protocol: 'https', hostname: '**' }]
-
-  if (isDevelopment()) {
-    patterns.push(...SAFE_LOCAL_IMAGE_REMOTE_PATTERNS)
-  }
-
-  return patterns
-}
 
 const nextConfig: NextConfig = {
   reactStrictMode: true,
